requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
boto==2.5.2
celery==3.1.18
Django==1.8.4
django-appconf==1.0.1
django-celery==3.1.16
django-crispy-forms==1.5.0
django-datetime-widget==0.9.3
django-debug-toolbar==1.3.2
django-filter==0.10.0
django-grappelli==2.6.5
django-storages==1.1.8
djangorestframework==3.1.2
djangorestframework-jwt==1.2.0
ecdsa==0.13
enum34==1.0.4
Fabric==1.4.3
httpie==0.9.2
idna==2.0
ipaddress==1.0.14
kombu==3.0.26
Markdown==2.6.2
paramiko==1.15.2
pilkit==1.1.12
Pillow==2.9.0
psycopg2==2.6.1
pyasn1==0.1.8
pycparser==2.14
pycrypto==2.6
Pygments==2.0.2
PyJWT==0.4.1
python-memcached==1.54
pytz==2015.4
redis==2.10.3
requests==2.7.0
six==1.9.0
sqlparse==0.1.16
ssh==1.7.14
